Fixed a little bug in mesh

+++ SDLGame-TFG/GraphicEngine/matrix4f.cpp
@@ -21,7 +21,7 @@
 
 Matrix4f::Matrix4f()
 {
-    matrix=new GLdouble[16];
+    matrix=new GLfloat[16];
 }
 
 //**********************************************************************//
@@ -33,8 +33,8 @@ Matrix4f::~Matrix4f()
 
 //**********************************************************************//
 
-void Matrix4f::translation(double x,double y,double z){
-    matrix[0]= 1; matrix[1]=0;  matrix[2]=0;  matrix[3]=0;
+void Matrix4f::translation(float x,float y,float z){
+    matrix[0]=1;  matrix[1]=0;  matrix[2]=0;  matrix[3]=0;
 	matrix[4]=0;  matrix[5]=1;  matrix[6]=0;  matrix[7]=0;
 	matrix[8]=0;  matrix[9]=0;  matrix[10]=1; matrix[11]=0;
 	matrix[12]=x; matrix[13]=y; matrix[14]=z; matrix[15]=1;
@@ -42,7 +42,7 @@ void Matrix4f::translation(double x,double y,double z){
 
 //**********************************************************************//
 
-void Matrix4f::scale(double x,double y,double z){
+void Matrix4f::scale(float x,float y,float z){
 
 	matrix[0]= x; matrix[1]=0;  matrix[2]=0;  matrix[3]=0;
 	matrix[4]=0;  matrix[5]=y;  matrix[6]=0;  matrix[7]=0;
@@ -58,8 +58,8 @@ void Matrix4f::rotation(float grade,const float x,const float y,const float z){
 
     if(x>0){
         matrix[0]= 1; matrix[1]=0;          matrix[2]=0;            matrix[3]=0;
-        matrix[4]=0;  matrix[5]=cos(rad);   matrix[6]=sin(rad);   matrix[7]=0;
-        matrix[8]=0;  matrix[9]=-sin(rad);  matrix[10]=cos(rad);  matrix[11]=0;
+        matrix[4]=0;  matrix[5]=cos(rad);   matrix[6]=sin(rad);     matrix[7]=0;
+        matrix[8]=0;  matrix[9]=-sin(rad);  matrix[10]=cos(rad);    matrix[11]=0;
         matrix[12]=0; matrix[13]=0;         matrix[14]=0;           matrix[15]=1;
     }
     else if(y>0){
@@ -80,6 +80,6 @@ void Matrix4f::rotation(float grade,const float x,const float y,const float z){
 
 //**********************************************************************//
 
-GLdouble * Matrix4f::getMatrix(){
+GLfloat * Matrix4f::getMatrix(){
     return matrix;
 }
